requirements.txt
Flask
spacy
scikit-learn
pdfminer.six
pandas
